Fix tests so they all pass now
This stopped working because it seems we didn't update this to include the `transliterated_word` key in the expected result. Also the word in expected at the root has to be a string key, not symbol to properly match.

Further, I've swapped the arguments for the assert_equal assertions since the first should be the expected while the second argument is that actual result.